+++ functions/bin/index.dart
@@ -3,9 +3,7 @@ library my_functions;
 import 'package:firebase_functions_interop/firebase_functions_interop.dart';
 
 void main() {
-  var functions = new FirebaseFunctions();
-
-  var httpsFunc = functions.https.onRequest((request, response) {
+  var httpsFunc = firebaseFunctions.https.onRequest((request, response) {
     String name = request.query['name'];
     if (name != null) {
       response.send('Hello to you $name from Dart Firebase Functions interop');
@@ -15,8 +13,9 @@ void main() {
   });
   exports.setProperty('helloWorld', httpsFunc);
 
-  var dbFunc =
-      functions.database.ref('/messages/{pushId}/original').onWrite((event) {
+  var dbFunc = firebaseFunctions.database
+      .ref('/messages/{pushId}/original')
+      .onWrite((event) {
     String original = event.data.val();
     print('Uppercasing $original');
     String uppercase = original.toUpperCase();